fix: review on custom-events.ts file
