+++ src/components/core/interfaces/types.ts
@@ -30,3 +30,9 @@ export interface SbbValueStateChange {
   type: 'value';
   value: string | null;
 }
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    stateChange: CustomEvent<SbbStateChange>;
+  }
+}
