+++ src/components/core/overlay/overlay.ts
@@ -122,3 +122,11 @@ export function removeInertMechanism(): void {
     removeSbbInert(el as HTMLElement),
   );
 }
+declare global {
+  interface GlobalEventHandlersEventMap {
+    willOpen: CustomEvent<void>;
+    willClose: CustomEvent<void>;
+    didOpen: CustomEvent<void>;
+    didClose: CustomEvent<void>;
+  }
+}
